+++ sdk/cognitivelanguage/azure-ai-language-conversations/azure/ai/language/conversations/authoring/aio/_client.py
@@ -7,31 +7,26 @@
 # --------------------------------------------------------------------------
 
 from copy import deepcopy
-from typing import Any, TYPE_CHECKING
+from typing import Any, Awaitable, TYPE_CHECKING
 
-from azure.core import PipelineClient
+from azure.core import AsyncPipelineClient
 from azure.core.credentials import AzureKeyCredential
-from azure.core.rest import HttpRequest, HttpResponse
+from azure.core.rest import AsyncHttpResponse, HttpRequest
 
-from ._configuration import ConversationAnalysisClientConfiguration
-from ._operations import ConversationAnalysisClientOperationsMixin
-from ._serialization import Deserializer, Serializer
+from .._serialization import Deserializer, Serializer
+from ._configuration import ConversationAuthoringClientConfiguration
+from ._operations import ConversationAuthoringClientOperationsMixin
 
 if TYPE_CHECKING:
     # pylint: disable=unused-import,ungrouped-imports
     from typing import Dict
 
-class ConversationAnalysisClient(ConversationAnalysisClientOperationsMixin):  # pylint: disable=client-accepts-api-version-keyword
-    """The language service conversations API is a suite of natural language processing (NLP) skills
-    that can be used to analyze structured conversations (textual or spoken). The synchronous API
-    in this suite accepts a request and mediates among multiple language projects, such as LUIS
-    Generally Available, Question Answering, Conversational Language Understanding, and then calls
-    the best candidate service to handle the request. At last, it returns a response with the
-    candidate service's response as a payload.
-
-     In some cases, this API needs to forward requests and responses between the caller and an
-    upstream service. The asynchronous APIs in this suite enable tasks like Conversation
-    Summarization and Conversational PII detection.
+class ConversationAuthoringClient(ConversationAuthoringClientOperationsMixin):  # pylint: disable=client-accepts-api-version-keyword
+    """The language service API is a suite of natural language processing (NLP) skills built with
+    best-in-class Microsoft machine learning algorithms. The API can be used to analyze
+    unstructured text for tasks such as sentiment analysis, key phrase extraction, language
+    detection and question answering. Further documentation can be found in :code:`<a
+    href="https://docs.microsoft.com/en-us/azure/cognitive-services/language-service/overview">https://docs.microsoft.com/en-us/azure/cognitive-services/language-service/overview</a>`.
 
     :param endpoint: Supported Cognitive Services endpoint (e.g.,
      https://:code:`<resource-name>`.cognitiveservices.azure.com). Required.
@@ -41,6 +36,8 @@ class ConversationAnalysisClient(ConversationAnalysisClientOperationsMixin):  # 
     :keyword api_version: Api Version. Default value is "2022-05-01". Note that overriding this
      default value may result in unsupported behavior.
     :paramtype api_version: str
+    :keyword int polling_interval: Default waiting time between two polls for LRO operations if no
+     Retry-After header is present.
     """
 
     def __init__(
@@ -50,8 +47,8 @@ class ConversationAnalysisClient(ConversationAnalysisClientOperationsMixin):  # 
         **kwargs: Any
     ) -> None:
         _endpoint = '{Endpoint}/language'
-        self._config = ConversationAnalysisClientConfiguration(endpoint=endpoint, credential=credential, **kwargs)
-        self._client = PipelineClient(base_url=_endpoint, config=self._config, **kwargs)
+        self._config = ConversationAuthoringClientConfiguration(endpoint=endpoint, credential=credential, **kwargs)
+        self._client = AsyncPipelineClient(base_url=_endpoint, config=self._config, **kwargs)
 
         self._serialize = Serializer()
         self._deserialize = Deserializer()
@@ -62,14 +59,14 @@ class ConversationAnalysisClient(ConversationAnalysisClientOperationsMixin):  # 
         self,
         request: HttpRequest,
         **kwargs: Any
-    ) -> HttpResponse:
+    ) -> Awaitable[AsyncHttpResponse]:
         """Runs the network request through the client's chained policies.
 
         >>> from azure.core.rest import HttpRequest
         >>> request = HttpRequest("GET", "https://www.example.org/")
         <HttpRequest [GET], url: 'https://www.example.org/'>
-        >>> response = client.send_request(request)
-        <HttpResponse: 200 OK>
+        >>> response = await client.send_request(request)
+        <AsyncHttpResponse: 200 OK>
 
         For more information on this code flow, see https://aka.ms/azsdk/dpcodegen/python/send_request
 
@@ -77,7 +74,7 @@ class ConversationAnalysisClient(ConversationAnalysisClientOperationsMixin):  # 
         :type request: ~azure.core.rest.HttpRequest
         :keyword bool stream: Whether the response payload will be streamed. Defaults to False.
         :return: The response of your network call. Does not do error handling on your response.
-        :rtype: ~azure.core.rest.HttpResponse
+        :rtype: ~azure.core.rest.AsyncHttpResponse
         """
 
         request_copy = deepcopy(request)
@@ -88,15 +85,12 @@ class ConversationAnalysisClient(ConversationAnalysisClientOperationsMixin):  # 
         request_copy.url = self._client.format_url(request_copy.url, **path_format_arguments)
         return self._client.send_request(request_copy, **kwargs)
 
-    def close(self):
-        # type: () -> None
-        self._client.close()
+    async def close(self) -> None:
+        await self._client.close()
 
-    def __enter__(self):
-        # type: () -> ConversationAnalysisClient
-        self._client.__enter__()
+    async def __aenter__(self) -> "ConversationAuthoringClient":
+        await self._client.__aenter__()
         return self
 
-    def __exit__(self, *exc_details):
-        # type: (Any) -> None
-        self._client.__exit__(*exc_details)
+    async def __aexit__(self, *exc_details) -> None:
+        await self._client.__aexit__(*exc_details)