+++ sdk/cognitivelanguage/azure-ai-language-conversations/azure/ai/language/conversations/authoring/aio/_configuration.py
@@ -15,8 +15,8 @@ from azure.core.pipeline import policies
 from .._version import VERSION
 
 
-class ConversationAnalysisClientConfiguration(Configuration):  # pylint: disable=too-many-instance-attributes
-    """Configuration for ConversationAnalysisClient.
+class ConversationAuthoringClientConfiguration(Configuration):  # pylint: disable=too-many-instance-attributes
+    """Configuration for ConversationAuthoringClient.
 
     Note that all parameters used to create this instance are saved as instance
     attributes.
@@ -37,7 +37,7 @@ class ConversationAnalysisClientConfiguration(Configuration):  # pylint: disable
         credential: AzureKeyCredential,
         **kwargs: Any
     ) -> None:
-        super(ConversationAnalysisClientConfiguration, self).__init__(**kwargs)
+        super(ConversationAuthoringClientConfiguration, self).__init__(**kwargs)
         api_version = kwargs.pop('api_version', "2022-05-01")  # type: str
 
         if endpoint is None: