+++ sdk/cognitivelanguage/azure-ai-language-conversations/CHANGELOG.md
@@ -10,6 +10,11 @@
 * Client now uses python dictionaries for method parameters and results instead of classes.
 * Many input and result parameter name changes in `analyze_conversation()` method
 
+## 1.0.0 (Unreleased)
+
+### Features Added
+* Added more resolution types for entities
+* Added support for authoring client
 
 ## 1.0.0b3 (2022-04-19)